Update and finish LD calcuations, refine population filtering to include merged refSNPs and duplicate ID removal

+++ src/calculate_ld.py
@@ -27,318 +27,10 @@ from . import log
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 
-def read_population_map(fp: str = globe._fp_population_map) -> pd.DataFrame:
-    """
-    Parse the 1KGP sample -> population map. The file associates individual sample IDs
-    with their population and superpopulation.
-
-    arguments
-        fp: filepath to the mapping
-
-    returns
-        a dataframe containing sample IDs and populations
-    """
-
-    return pd.read_csv(fp, sep='\t')[['sample', 'pop', 'super_pop']]
-
-
-def show_population_groups(df: pd.DataFrame) -> None:
-    """
-    Display population and superpopulation groups.
-
-    arguments
-        df: dataframe containing sample -> population map
-    """
-
-    table = df[['pop', 'super_pop']].drop_duplicates(subset='pop')
-    table = table.sort_values(by=['super_pop', 'pop']).set_index('super_pop')
-
-    print(table)
-
-
-def extract_header(fp: str) -> List[str]:
-    """
-    Find and attempt to extract the header from the given VCF file.
-
-    arguments
-        fp: filepath to the VCF file
-
-    returns
-        a list of header fields
-    """
-
-    header = []
-
-    with open(fp, 'r') as fl:
-        for ln in fl:
-            if not ln:
-                continue
-
-            ## Field rows should be prefix by '##'
-            if ln[:2] == '##':
-                continue
-
-            ## The header row should be prefix by a single '#'
-            if ln[0] == '#':
-                header = ln[1:].strip().split('\t')
-
-            break
-
-    if not header:
-        raise exceptions.VCFHeaderParsingFailed(
-            f'Could not discern VCF header from {fp}'
-        )
-
-    return header
-
-
-def retain_samples(pops: List[str], popmap: pd.DataFrame, super_pop=False) -> List[str]:
-    """
-    Determine what variant call samples to keep based on population structure.
-    Returns a list of samples that should be EXCLUDED from further analysis.
-
-    arguments
-        pops:      a list of population IDs
-        popmap:    the sample -> population mapping dataframe
-        super_pop: if True, the population list contains a list of super population IDs
-
-    returns
-        a list of samples to retain
-    """
-
-    ## Get sample IDs associated with the given population
-    if super_pop:
-        samples = popmap[popmap.super_pop.isin(pops)]['sample']
-    else:
-        samples = popmap[popmap['pop'].isin(pops)]['sample']
-
-    ## Now get all other samples that should be excluded
-    samples = popmap[~popmap['sample'].isin(samples)]['sample']
-
-    return samples.tolist()
-
-
-def read_vcf(fp: str):
-    """
-    Read in a VCF file.
-    """
-
-    header = extract_header(fp)
-
-    return ddf.read_csv(fp, sep='\t', comment='#', header=None, names=header)
-
-
-def filter_vcf_populations(
-        fp: str,
-        pops: List[str],
-        super_pop: bool = False
-) -> pd.DataFrame:
-    """
-    Filter variant call samples based on population structure.
-
-    arguments
-        df:
-        pops:
-
-    returns
-    """
-
-    df = read_vcf(fp)
-    popmap = read_population_map()
-    samps = retain_samples(pops, popmap, super_pop=super_pop)
-
-    return df.drop(labels=samps, axis=1)
-
-
-def filter_vcf_populations_serial(
-        fp: str,
-        pops: List[str],
-        super_pop: bool = False
-) -> pd.DataFrame:
-    """
-    Filter variant call samples based on population structure.
-
-    arguments
-        df:
-        pops:
-
-    returns
-    """
-
-    popmap = read_population_map()
-    samps = retain_samples(pops, popmap, super_pop=super_pop)
-    header = extract_header(fp)
-
-    for vcf in pd.read_csv(
-            fp,
-            sep='\t',
-            header=None,
-            names=header,
-            comment='#',
-            chunksize=14096
-    ):
-
-        ## Filter out samples not in our selected populations
-        #vcf = vcf.filter(items=samps)
-        vcf.drop(labels=samps, axis=1)
-
-        ## Save as output
-        vcf.to_csv('data/shit.vcf', sep='\t', header=False, index=False, mode='a')
-
-
-def write_dataframe(df: pd.DataFrame, fp: str, first: bool = False, **kwargs) -> None:
-    """
-
-    arguments
-        df:
-        fp:
-        first:
-
-    returns
-    """
-
-    mode = 'w' if first else 'a'
-
-    df.to_csv(fp, sep='\t', index=False, header=first, mode=mode)
-
-
-def save_variant_dataframe(df: ddf.DataFrame, fp: str) -> None:
-    """
-
-    arguments
-        df:
-    """
-
-    dels = df.to_delayed()
-
-    log._logger.info(f'Saving variants to {fp}')
-
-    last_delayed = dask.delayed(write_dataframe)(dels[0], fp, first=True)
-
-    for d in dels[1:]:
-        last_delayed = dask.delayed(write_dataframe)(
-            d, fp, first=False, depends=last_delayed
-        )
-
-    return last_delayed
-
-
-def save_variant_dataframe2(df: ddf.DataFrame, out_fp: str) -> None:
-    """
-
-    arguments
-        df:
-    """
-
-    log._logger.info(f'Saving variants to {out_fp}')
-    log._logger.info(df.head())
-
-    ## Write all dataframe partitions to a temp dir
-    #with tf.TemporaryDirectory(dir=globe._dir_1k_processed) as tmpdir:
-    tmpdir = tf.mkdtemp(dir=globe._dir_1k_processed)
-
-    df.to_csv(tmpdir, sep='\t', index=False)
-
-    log._logger.info(f'Finished writing the variants to {tmpdir}')
-
-    first = True
-
-    ## Concat all frames into a single file
-    with open(out_fp, 'w') as ofl:
-
-        for tmpfp in Path(tmpdir).iterdir():
-
-            log._logger.info(f'Reading/writing {tmpfp}')
-            with open(tmpfp, 'r') as tmpfl:
-
-                ## If it's the first file we include the header
-                if first:
-                    ofl.write(tmpfl.read())
-
-                ## Otherwise skip the header so it isn't repeated
-                else:
-                    next(tmpfl)
-
-                    ofl.write(tmpfl.read())
-
-                first = False
-
-    ## Remove the temp directory we made and it's contents
-    shutil.rmtree(tmpdir)
-
-    return True
-
-
-"""
-def filter_populations(client: Client, pops: List[str], super_pop: bool = False):
-
-    ## List of delayed objects that will be written to files
-    saved = []
-
-    for chrom in range(1, 7):
-        vcf_path = globe._fp_1k_variant_autosome % chrom
-        out_path = globe._fp_1k_processed_autosome % chrom
-
-        df = filter_vcf_populations(vcf_path, pops, super_pop=super_pop)
-        df_save = save_variant_dataframe(df, out_path)
-
-        saved.append(df_save)
-
-    #x_df = filter_vcf_populations(globe._fp_1k_variant_x, pops, super_pop=super_pop)
-    #x_df_save = save_variant_dataframe(x_df, globe._fp_1k_processed_x)
-
-    #y_df = filter_vcf_populations(globe._fp_1k_variant_y, pops, super_pop=super_pop)
-    #y_df = save_variant_dataframe(y_df, globe._fp_1k_processed_y)
-
-    #saved.append(x_df_save)
-    #saved.append(y_df)
-
-    ## Wait for computations, i.e., file writing as delayed objects, to finish
-    #client.compute(saved)
-    return saved
-"""
-
-def filter_populations2(client: Client, pops: List[str], super_pop: bool = False):
-    """
-    """
-
-    ## List of delayed objects that will be written to files
-    saved = []
-
-    #for chrom in range(21, 22):
-    for chrom in range(1, 10):
-        vcf_path = globe._fp_1k_variant_autosome % chrom
-        out_path = globe._fp_1k_processed_autosome % chrom
-
-        df = client.submit(filter_vcf_populations, vcf_path, pops, super_pop=super_pop)
-        #df = filter_vcf_populations(vcf_path, pops, super_pop=super_pop)
-        #dfs = client.scatter(df)
-        #df = filter_vcf_populations(vcf_path, pops, super_pop=super_pop)
-        #future = save_variant_dataframe(df, out_path)
-        future = client.submit(save_variant_dataframe2, df, out_path)
-        #future = client.submit(save_variant_dataframe2, dfs, out_path)
-
-        saved.append(future)
-
-    #x_df = filter_vcf_populations(globe._fp_1k_variant_x, pops, super_pop=super_pop)
-    #x_df = save_variant_dataframe(x_df, globe._fp_1k_processed_x)
-
-    #y_df = filter_vcf_populations(globe._fp_1k_variant_y, pops, super_pop=super_pop)
-    #y_df = save_variant_dataframe(y_df, globe._fp_1k_processed_y)
-
-    #saved.append(x_df)
-    #saved.append(y_df)
-
-    ## Wait for computations, i.e., file writing as delayed objects, to finish
-    #client.compute(saved)
-    return saved
-
-
 def _calculate_ld(
     input: str,
     output: str,
-    piece: int = 1,
-    pieces: int = 1,
+    snps: str,
     r2: float = 0.7,
     threads: int = 3,
     plink: str = globe._exe_plink
@@ -358,13 +50,10 @@ def _calculate_ld(
             plink,
             '--threads',
             str(threads),
-            '--memory',
-            '40000',
-            '--parallel',
-            str(piece),
-            str(pieces),
             '--vcf',
             input,
+            '--ld-snp-list',
+            snps,
             '--ld-window-r2',
             str(r2),
             '--r2',
@@ -377,13 +66,14 @@ def _calculate_ld(
         log._logger.error('There was a problem running plink as a subprocess: %s', e)
         raise
 
-    return True
+    ## Plink auto appends an '.ld' suffix to the output which we add here
+    return Path(output).with_suffix(f'{Path(output).suffix}.ld')
 
 
 def calculate_ld(
     vcf_dir: str,
-    out_dir: str,
-    pieces: int = 6,
+    snps: str,
+    out_dir: str = globe._dir_work,
     r2: float = 0.7,
     threads: int = 3
 ):
@@ -403,62 +93,55 @@ def calculate_ld(
     ## Iterate through all the VCF files
     for fp in Path(vcf_dir).iterdir():
 
-        ## If there are no more workers left in the list (means there are more than 23
-        ## VCF files in the directory) then we can't do anything
-        #if not workers:
-        #    log._logger.error(
-        #        f'No workers are left for calculating LD, skipping {fp.as_posix()}'
-        #    )
-
         ## Construct the output path
-        output = Path(out_dir, fp.stem)
+        output = Path(out_dir, '-'.join([fp.stem, Path(snps).stem]))
 
+        ## Get a worker
+        #worker = workers.pop()
+        ### Send him to the back of the worker queue
+        #workers.append(worker)
 
         ## Calculate LD on the worker
-        #future = client.submit(
-        #    _calculate_ld, fp.as_posix(), output.as_posix(), workers=[worker]
-        #)
-        ## Break calculation into pieces for each worker
-        for p in range(1, pieces + 1):
+        future = client.submit(
+            _calculate_ld,
+            fp.as_posix(),
+            output.as_posix(),
+            snps,
+            r2=r2,
+            threads=threads,
+            #workers=[worker]
+        )
 
-            ## Get a worker
-            worker = workers.pop()
-            ## Send him to the back of the worker queue
-            workers.append(worker)
-
-            future = client.submit(
-                _calculate_ld,
-                fp.as_posix(),
-                output.as_posix(),
-                piece=p,
-                pieces=pieces,
-                r2=r2,
-                threads=threads,
-                workers=[worker]
-            )
-
-            futures.append(future)
+        futures.append(future)
 
     return futures
 
 
-def _format_merge_files(files: List[str]):
+def format_merge_files(files: List[str], output: str, delete: bool = True) -> str:
     """
     """
 
     if not files:
-        return ''
+        return output
 
-    ## Make a single output path
-    output = Path(files[0])
-    output = Path(output.parent, output.stem).as_posix()
+    first = True
 
     ## Open the single concatenated output file
     with open(output, 'w') as outfl:
+
         ## Loop through input files...
         for fpath in sorted(files):
+
             ## Read each input file and format line x line
             with open(fpath, 'r') as infl:
+
+                if not first:
+                    ## Skip the header
+                    next(infl)
+
+                else:
+                    first = False
+
                 for ln in infl:
                     ln = re.sub(r'^ +', '', ln)
                     ln = re.sub(r' +', '\t', ln)
@@ -466,7 +149,10 @@ def _format_merge_files(files: List[str]):
                     outfl.write(ln)
 
             ## Remove the file once we're done
-            Path(fpath).unlink()
+            if delete:
+                Path(fpath).unlink()
+
+    return output
 
 
 def format_merge_separate_files(out_dir: str = globe._dir_1k_ld):
